Ok, so I've benchmarked again everything and all is green for the zircon saturation. 
I've also slighty increase the performance without pushing too much (there is a lot of array allocations, but there is no easy way to get ride of it with how the code is structured, and as the input oxides depends on the database)

Before this PR:

```
# for Boehnke et al., 2013
# 2.136 μs (35 allocations: 3.19 KiB) 
# Crisp and Berry 2022
# 3.480 μs (62 allocations: 6.16 KiB) CB
```

After this PR:

```
# for Boehnke et al., 2013
# 1.409 μs (18 allocations: 2.12 KiB)
# Crisp and Berry 2022
# 2.592 μs (29 allocations: 3.94 KiB)
```